Add files via upload
Table 1 Capacities (cubic meters of roundwood) and number of facilities per country. For pulp and paper , the list does not distinguish between raw and recycled materials. For bioenergy, the list reports the capacities of industrial bioenergy plants using woody biomass as feedstock. The capacity of sawmills is not available for Switzerland and France.